pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codeverse-cli"
version = "1.0.0"
authors = [
    {name = "IBDA AI LTD", email = "support@ibda.me"},
]
description = "AI-powered development assistant with 54 specialized agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Environment :: Console",
    "Operating System :: OS Independent",
]
keywords = ["ai", "cli", "development", "assistant", "code-generation", "automation"]
dependencies = [
    "requests>=2.31.0",
    "websocket-client>=1.6.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "cryptography>=41.0.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.8.0",
    "aiofiles>=23.0.0",
    "watchdog>=3.0.0",
    "gitpython>=3.1.0",
    "keyring>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
"Homepage" = "https://codeverse.ibda.me"
"Documentation" = "https://docs.codeverse.ibda.me"
"Repository" = "https://github.com/ibda-ai/codeverse-cli"
"Bug Tracker" = "https://github.com/ibda-ai/codeverse-cli/issues"
"Changelog" = "https://github.com/ibda-ai/codeverse-cli/blob/main/CHANGELOG.md"

[project.scripts]
codeverse = "codeverse_cli.main:main"
cv = "codeverse_cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["codeverse_cli*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
codeverse_cli = ["*.json", "*.yaml", "*.yml"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]